refactor game engine factories to individual folders
