+++ TheAlienThatEatsTheCarrot/GameEngine/Entities/Factories/EnemyFactories/NormalEnemyFactory.swift
@@ -1,0 +1,18 @@
+//
+//  NormalEnemyFactory.swift
+//  TheAlienThatEatsTheCarrot
+//
+//  Created by Justin Cheah Yun Fei on 29/3/24.
+//
+
+import Foundation
+
+class NormalEnemyFactory: EnemyFactory {
+    override func createComponents() -> [Component] {
+        let enemyComponent = EnemyComponent(entity: entity)
+        let renderableComponent = RenderableComponent(entity: entity, position: boardObject.position, objectType: .enemy(.normal))
+        let movableComponent = MovableComponent(entity: entity, velocity: CGVector(dx: 50.0, dy: 0), movementPattern: .leftRight)
+        let destroyableComponent = DestroyableComponent(entity: entity)
+        return [enemyComponent, renderableComponent, movableComponent, destroyableComponent]
+    }
+}
